+++ CHANGELOG.md
@@ -4,6 +4,7 @@
     - Remove inline `substrate#AxisLayer` declaration
     - Remove reliance upon downloaded CDN three.js
     - Enable figsize specification during construction
+    - Add debug mode with `debug=True` constructor arg
 - **0.4.1** Unreleased
     - Resolve integer-label graph conflict (#80)
 - **0.4.0** (July 31, 2018)